+++ Optuna/Dashboard/HumanInTheLoop/HumanSliderInput.cs
@@ -1,7 +1,5 @@
 using System.Collections.Generic;
 using System.Drawing;
-using System.Globalization;
-using System.Linq;
 using System.Text;
 
 using Optuna.Study;
@@ -14,7 +12,6 @@ namespace Optuna.Dashboard.HumanInTheLoop
     public class HumanSliderInput : HumanInTheLoopBase
     {
         private readonly PyModule _importedLibrary;
-        private readonly dynamic _artifactBackend;
 
         public HumanSliderInput(string tmpPath, string storagePath) : base(tmpPath, storagePath)
         {
@@ -24,18 +21,13 @@ namespace Optuna.Dashboard.HumanInTheLoop
             importedLibrary.Exec("from optuna_dashboard import ObjectiveUserAttrRef");
             importedLibrary.Exec("from optuna_dashboard import register_objective_form_widgets");
             importedLibrary.Exec("from optuna_dashboard import set_objective_names");
-            importedLibrary.Exec("from optuna_dashboard.artifact import upload_artifact");
-            importedLibrary.Exec("from optuna_dashboard.artifact.file_system import FileSystemBackend");
-
-            dynamic fileSystemBackend = importedLibrary.Get("FileSystemBackend");
-            _artifactBackend = fileSystemBackend(base_path: _artifactPath);
             _importedLibrary = importedLibrary;
         }
 
         public void SetObjective(StudyWrapper study, string[] objectiveNames)
         {
             dynamic setObjectiveNames = _importedLibrary.Get("set_objective_names");
-            PyList pyNameList = EnumeratorToPyList(objectiveNames.Select(s => s.Replace("Human-in-the-Loop", "HITL")));
+            PyList pyNameList = EnumeratorToPyList(objectiveNames);
             setObjectiveNames(study.PyInstance, pyNameList);
         }
 
@@ -50,7 +42,7 @@ namespace Optuna.Dashboard.HumanInTheLoop
             {
                 if (objectiveNames[i].Contains("Human-in-the-Loop"))
                 {
-                    widgets[i] = sliderWidget(min: 1, max: 5, step: 1, description: "Smaller is better");
+                    widgets[i] = sliderWidget(min: 1, max: 5, step: 1, description: "Your rating of the image: Smaller is better👍");
                 }
                 else
                 {
@@ -64,18 +56,16 @@ namespace Optuna.Dashboard.HumanInTheLoop
 
         public void SaveNote(StudyWrapper study, TrialWrapper trial, Bitmap[] bitmaps)
         {
-            dynamic uploadArtifact = _importedLibrary.Get("upload_artifact");
             var noteText = new StringBuilder();
             noteText.AppendLine("# Image");
             noteText.AppendLine("");
 
-            CheckDirectoryIsExist();
             for (int i = 0; i < bitmaps.Length; i++)
             {
                 Bitmap bitmap = bitmaps[i];
-                string path = $"{_tmpPath}/image_{study.Id}_{trial.Id}.png";
-                bitmap?.Save(path, System.Drawing.Imaging.ImageFormat.Png);
-                dynamic artifactId = uploadArtifact(_artifactBackend, trial.PyInstance, path);
+                string filePath = $"{_tmpPath}/image_{trial.Number}.png";
+                bitmap?.Save(filePath, System.Drawing.Imaging.ImageFormat.Png);
+                string artifactId = ArtifactStore.UploadArtifact(filePath, trial);
                 string str = $"![](/artifacts/{study.Id}/{trial.Id}/{artifactId})";
                 noteText.AppendLine(str);
             }