+++ Tunny/WPF/ViewModels/Optimize/OptimizeViewModel.cs
@@ -22,7 +22,6 @@ using Tunny.Core.Util;
 using Tunny.Eto.Common;
 using Tunny.Eto.Message;
 using Tunny.Eto.Models;
-using Tunny.Process;
 using Tunny.WPF.Common;
 using Tunny.WPF.Models;
 using Tunny.WPF.Views.Pages.Optimize;
@@ -43,6 +42,7 @@ namespace Tunny.WPF.ViewModels.Optimize
         private Lazy<TPESettingsPage> _tpePage;
         private Lazy<GPOptunaSettingsPage> _gpOptunaPage;
         private Lazy<GPBoTorchSettingsPage> _gpBoTorchPage;
+        private Lazy<PreferentialGpSettingsPage> _preferentialGpPage;
         private Lazy<NSGAIISettingsPage> _nsgaiiPage;
         private Lazy<NSGAIIISettingsPage> _nsgaiiiPage;
         private Lazy<CmaEsSettingsPage> _cmaesPage;
@@ -229,6 +229,7 @@ namespace Tunny.WPF.ViewModels.Optimize
             _moeadPage = new Lazy<MOEADSettingsPage>(() => MOEADSettingsPage.FromSettings(_settings));
             _moCmaEsPage = new Lazy<MoCmaEsSettingsPage>(() => MoCmaEsSettingsPage.FromSettings(_settings));
             _dePage = new Lazy<DESettingsPage>(() => DESettingsPage.FromSettings(_settings));
+            _preferentialGpPage = new Lazy<PreferentialGpSettingsPage>(() => PreferentialGpSettingsPage.FromSettings(_settings));
         }
 
         public void ChangeTargetSampler(SamplerType samplerType)
@@ -288,6 +289,10 @@ namespace Tunny.WPF.ViewModels.Optimize
                     param = _dePage.Value;
                     OptimizeSettingsPage = _dePage.Value;
                     break;
+                case SamplerType.PreferentialGp:
+                    param = _preferentialGpPage.Value;
+                    OptimizeSettingsPage = _preferentialGpPage.Value;
+                    break;
                 default:
                     throw new ArgumentOutOfRangeException(nameof(samplerType), samplerType, null);
             }
@@ -305,6 +310,10 @@ namespace Tunny.WPF.ViewModels.Optimize
                 int numGeneration = total / populationSize;
                 TrialNumberParam1 = numGeneration.ToString(CultureInfo.InvariantCulture);
                 TrialNumberParam2 = populationSize.ToString(CultureInfo.InvariantCulture);
+            }
+            else if (CoSharedItems.Component.GhInOut.IsHumanInTheLoop)
+            {
+                TrialNumberParam1 = "6";
             }
             else
             {